+++ conftest.py
@@ -1,6 +1,10 @@
 from unittest.mock import patch
 
 import pytest
+from wagtail.wagtailimages.models import Image
+
+from django.core.files.storage import default_storage
+from django.core.files.uploadedfile import SimpleUploadedFile
 
 
 @pytest.fixture(autouse=True)
@@ -24,3 +28,24 @@ def mock_auth():
     stub.start()
     yield stub
     stub.stop()
+
+
+@pytest.fixture
+def uploaded_file():
+    return SimpleUploadedFile(
+        name='test_image.png',
+        content=open('core/static/core/logo.png', 'rb').read(),
+        content_type='image/png'
+    )
+
+
+@pytest.fixture
+def image(uploaded_file):
+    image = Image.objects.create(
+        file=uploaded_file,
+        title='test',
+        width=100,
+        height=100,
+    )
+    yield image
+    default_storage.delete(image.file.name)
