+++ requirements.txt
@@ -11,6 +11,8 @@
 -e git+https://github.com/uktrade/export-elements.git@v0.3.0#egg=export_elements
 -e git+https://github.com/uktrade/directory-signature-auth.git@v2.0.0#egg=sigauth
 beautifulsoup4==4.6.0     # via wagtail
+boto3==1.6.3
+botocore==1.9.3           # via boto3, s3transfer
 cachetools==2.0.1         # via google-auth
 certifi==2018.1.18        # via requests
 chardet==3.0.4            # via requests
@@ -25,6 +27,7 @@ django-treebeard==4.2.0   # via wagtail
 django==1.11.5
 django_storages==1.6.5
 djangorestframework==3.4.7  # via wagtail
+docutils==0.14            # via botocore
 factory-boy==2.10.0       # via wagtail-factories
 faker==0.8.11             # via factory-boy
 gevent==1.2.2
@@ -37,6 +40,7 @@ greenlet==0.4.13          # via gevent
 gunicorn==19.5.0
 html5lib==0.999999999     # via wagtail
 idna==2.6                 # via requests
+jmespath==0.9.3           # via boto3, botocore
 mohawk==0.3.4
 olefile==0.45.1           # via pillow
 pillow==4.3.0             # via wagtail
@@ -44,11 +48,12 @@ protobuf==3.5.1           # via google-api-core, googleapis-common-protos
 psycopg2==2.7.3.2
 pyasn1-modules==0.2.1     # via google-auth
 pyasn1==0.4.2             # via pyasn1-modules, rsa
-python-dateutil==2.6.1    # via faker
+python-dateutil==2.6.1    # via botocore, faker
 pytz==2017.2              # via django, django-modelcluster, google-api-core
 raven==5.19.0
 requests==2.18.4          # via google-api-core, wagtail
 rsa==3.4.2                # via google-auth
+s3transfer==0.1.13        # via boto3
 setuptools-git==1.2
 six==1.11.0               # via faker, google-api-core, google-auth, html5lib, mohawk, protobuf, python-dateutil
 text-unidecode==1.1       # via faker
