+++ requirements_test.txt
@@ -11,6 +11,8 @@
 -e git+https://github.com/uktrade/export-elements.git@v0.3.0#egg=export_elements
 -e git+https://github.com/uktrade/directory-signature-auth.git@v2.0.0#egg=sigauth
 beautifulsoup4==4.6.0
+boto3==1.6.3
+botocore==1.9.3           # via boto3, s3transfer
 cachetools==2.0.1         # via google-auth
 certifi==2018.1.18        # via requests
 chardet==3.0.4            # via requests
@@ -27,6 +29,7 @@ django-treebeard==4.2.0   # via wagtail
 django==1.11.5
 django_storages==1.6.5
 djangorestframework==3.4.7  # via wagtail
+docutils==0.14            # via botocore
 factory-boy==2.8.1
 faker==0.8.11             # via factory-boy
 flake8==3.0.4
@@ -41,6 +44,7 @@ greenlet==0.4.13          # via gevent
 gunicorn==19.5.0
 html5lib==0.999999999     # via wagtail
 idna==2.6                 # via requests
+jmespath==0.9.3           # via boto3, botocore
 mccabe==0.5.3             # via flake8
 mohawk==0.3.4
 olefile==0.45.1           # via pillow
@@ -57,12 +61,13 @@ pytest-cov==2.3.1
 pytest-django==3.1.2
 pytest-sugar==0.8.0
 pytest==3.0.2
-python-dateutil==2.6.1    # via faker, freezegun
+python-dateutil==2.6.1    # via botocore, faker, freezegun
 pytz==2017.2              # via django, django-modelcluster, google-api-core
 raven==5.19.0
 requests==2.18.4          # via codecov, google-api-core, requests-mock, wagtail
 requests_mock==1.1.0
 rsa==3.4.2                # via google-auth
+s3transfer==0.1.13        # via boto3
 setuptools-git==1.2
 six==1.11.0               # via faker, freezegun, google-api-core, google-auth, html5lib, mohawk, protobuf, python-dateutil, requests-mock
 termcolor==1.1.0          # via pytest-sugar
